refactor: removed usePrimaryAccount
