+++ apps/extension/src/ui/apps/popup/pages/Login.tsx
@@ -1,4 +1,4 @@
-import { AccountJsonAny } from "@core/domains/accounts/types"
+import { AccountJsonAny, AccountType, storedSeedAccountTypes } from "@core/domains/accounts/types"
 import { yupResolver } from "@hookform/resolvers/yup"
 import { SuspenseTracker } from "@talisman/components/SuspenseTracker"
 import { HandMonoTransparentLogo } from "@talisman/theme/logos"
@@ -6,10 +6,10 @@ import { useTalismanOrb } from "@talismn/orb"
 import { classNames } from "@talismn/util"
 import { api } from "@ui/api"
 import { useAnalytics } from "@ui/hooks/useAnalytics"
-import { usePrimaryAccount } from "@ui/hooks/usePrimaryAccount"
-import { Suspense, useCallback, useEffect, useRef, useState } from "react"
+import { Suspense, useCallback, useEffect, useMemo, useRef, useState } from "react"
 import { SubmitHandler, useForm } from "react-hook-form"
 import { useTranslation } from "react-i18next"
+import { atom, useRecoilValue } from "recoil"
 import { Button, FormFieldInputText } from "talisman-ui"
 import { LoginBackground } from "talisman-ui"
 import * as yup from "yup"
@@ -35,6 +35,34 @@ const useAccountColors = ({ account }: { account: AccountJsonAny }): [string, st
   return [bgColor1, bgColor2]
 }
 
+// do not use this atom outside of this screen as it may cause performance issues because of suspense
+const accountsState = atom<AccountJsonAny[]>({
+  key: "accountsState",
+  effects: [
+    ({ setSelf }) => {
+      const unsubscribe = api.accountsSubscribe(setSelf)
+      return () => unsubscribe()
+    },
+  ],
+})
+
+/**
+ * @param storedSeedOnly causes the function to return only accounts derived from a seed stored in Talisman. Returns that account if it
+ * exists, or the first account present if not, by default
+ * @returns an Account
+ */
+export const useBackgroundLoginAccount = (storedSeedOnly?: boolean) => {
+  const accounts = useRecoilValue(accountsState)
+
+  const storedSeedAccount = useMemo(
+    () => accounts.find(({ origin }) => storedSeedAccountTypes.includes(origin as AccountType)),
+    [accounts]
+  )
+  if (storedSeedOnly) return storedSeedAccount
+  if (accounts.length > 0) return accounts[0]
+  return
+}
+
 const LoginBackgroundWithAccount = ({ account }: { account: AccountJsonAny }) => {
   const accountColors = useAccountColors({ account })
 
@@ -49,10 +77,10 @@ const LoginBackgroundWithAccount = ({ account }: { account: AccountJsonAny }) =>
 }
 
 const Background = () => {
-  const primaryAccount = usePrimaryAccount()
+  const account = useBackgroundLoginAccount()
 
-  return primaryAccount ? (
-    <LoginBackgroundWithAccount account={primaryAccount} />
+  return account ? (
+    <LoginBackgroundWithAccount account={account} />
   ) : (
     <LoginBackground
       width={400}
